Eliminado .to en proposals

+++ CustomFaster_MP_TPU/generalized_rcnn.py
@@ -122,7 +122,7 @@ class GeneralizedRCNN(nn.Module):
         if isinstance(features, torch.Tensor):
             features = OrderedDict([("0", features)])
         proposals, proposal_losses = self.rpn(new_images, features, targets)
-        detections, detector_losses = self.roi_heads(features, proposals.to(torch.bfloat16), new_images.image_sizes, targets)
+        detections, detector_losses = self.roi_heads(features, proposals, new_images.image_sizes, targets)
         
         #bloque modificado--> no se hace postprocess aquí porque las imágenes ya está redimensionadas
         #detections = self.transform.postprocess(detections, new_images.size(), original_image_sizes)  # type: ignore[operator]
